# openapi-appengine.yaml
swagger: "2.0"
...
info:
  description: "Dantion Endpoints API."
  title: "Endpoints Dantion"
  version: "1.0.0"
host: "danger-detection.appspot.com"
...
consumes:
  - "application/json"
produces:
  - "application/json"
schemes:
  - "https"
  - "http"

# security:
#   - api_key: []
securityDefinitions:
  # api_key:
  #   type: "apiKey"
  #   name: "key"
  #   in: "query"
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
paths:
...
  "/admins/userRole":
    patch:
      tags:
        - "Admin"
      security:
        - Bearer: []
      description: "Update User Role"
      operationId: "UserRole"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: Admin Update User Role
          description: Request body for update user role.
          schema:
            $ref: "#/definitions/AdminUpdateUserRoleBody"
      responses:
        201:
          description: "Success update user role"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
...
  "/detections":
    get:
      tags:
        - "Detections"
      security:
        - Bearer: []
      description: "Get All Detection"
      operationId: "alldetection"
      produces:
        - "application/json"
      responses:
        201:
          description: "success get all detection"
          schema:
            $ref: "#/definitions/DetectionAllSuccessResponse"
        400:
          description: "failed get all detection"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
    post:
      tags:
        - "Detections"
      security:
        - Bearer: []
      description: "Add Detection"
      operationId: "adddetection"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: Add detection
          description: Request body for add detection.
          schema:
            $ref: "#/definitions/AddDetectionBody"
      responses:
        201:
          description: "success add detection"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
        400:
          description: "failed login user"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
    patch:
      tags:
        - "Detections"
      security:
        - Bearer: []
      description: "Update Detection"
      operationId: "updatedetection"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: Update detection
          description: Request body for Update detection.
          schema:
            $ref: "#/definitions/DetectionUpdateBody"
      responses:
        201:
          description: "success update detection"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
        400:
          description: "failed update detection"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
  "/detections/{id}":
    get:
      tags:
        - "Detections"
      security:
        - Bearer: []
      description: "Get Detection Using ID"
      operationId: "detectionId"
      produces:
        - "application/json"
      parameters:
        - name: id
          in: path
          description: "ID of detection"
          required: true
          type: "string"
      responses:
        201:
          description: "success get detail detection"
          schema:
            $ref: "#/definitions/DetectionDetailSuccessResponse"
        400:
          description: "failed get detail detection"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
    delete:
      tags:
        - "Detections"
      security:
        - Bearer: []
      description: "Delete Detection ID"
      operationId: "deleteDetection"
      produces:
        - "application/json"
      parameters:
        - name: id
          in: path
          description: "ID of detection"
          required: true
          type: "string"
      responses:
        201:
          description: "success delete detection"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
        400:
          description: "failed delete detection"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
...
definitions:
  StatusMessageResponse:
    type: "object"
    properties:
      error:
        type: "boolean"
      message:
        type: "string"
...
  AdminUpdateUserRoleBody:
    type: "object"
    properties:
      userId:
        type: "string"
      role:
        type: "string"
        enum:
          - "polisi"
          - "ambulan"
          - "damkar"
...
  DetectionDetailSuccessResponse:
    type: "object"
    properties:
      lat:
        type: "number"
        format: "double"
      lon:
        type: "number"
        format: "double"
      recordUrl:
        type: "string"
        format: "binary"
      type:
        type: "string"
        enum:
          - "kejahatan"
          - "kecelakaan"
          - "kebakaran"
      status:
        type: "string"
        enum:
          - "invalid"
          - "valid"
          - "selesai"
          - "tolak"
      city:
        type: "string"
      updatedAt:
        type: "string"
        format: "date-time"
      name:
        type: "string"
      address:
        type: "string"
      number:
        type: "string"
      parentNumber:
        type: "string"
  DetectionAllSuccessResponse:
    type: "object"
    properties:
      lat:
        type: "number"
        format: "double"
      lon:
        type: "number"
        format: "double"
      recordUrl:
        type: "string"
        format: "binary"
      type:
        type: "string"
        enum:
          - "kejahatan"
          - "kecelakaan"
          - "kebakaran"
      status:
        type: "string"
        enum:
          - "invalid"
          - "valid"
          - "selesai"
          - "tolak"
      city:
        type: "string"
      updatedAt:
        type: "string"
        format: "date-time"
      name:
        type: "string"
  AddDetectionBody:
    type: "object"
    properties:
      lat:
        type: "number"
        format: "double"
      lon:
        type: "number"
        format: "double"
      type:
        type: "string"
        enum:
          - "kejahatan"
          - "kecelakaan"
          - "kebakaran"
      userId:
        type: "string"
      city:
        type: "string"
      recordUrl:
        type: "string"
        format: "binary"
  DetectionUpdateBody:
    type: "object"
    properties:
      id:
        type: "string"
      status:
        type: "boolean"
        default: "invalid"
        enum:
          - "invalid"
          - "selesai"
          - "tolak"
          - "valid"
      idUserLogin:
        type: "string"
...